Completed Statistics and User API (fix and update)
